requirements.txt
opencv-python
pillow
numpy
torch
open_clip_torch
faiss-cpu
streamlit